requirements.txt
astroid==2.2.5
Babel==2.9.0
blis==0.2.4
certifi==2019.3.9
chardet==3.0.4
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
Click==7.0
colorama==0.4.1
cymem==2.0.2
dnspython==2.0.0
en==0.0.1
Flask==1.0.2
Flask-Babel==2.0.0
Flask-Login==0.5.0
Flask-PyMongo==2.3.0
Flask-SocketIO==4.3.1
Flask-Table==0.5.0
idna==2.8
importlib-metadata==2.0.0
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.17.0
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mathparse==0.1.2
mccabe==0.6.1
murmurhash==1.0.2
nltk==3.4
numpy==1.16.2
packaging==20.4
Pillow==8.0.1
Pint==0.9
plac==0.9.6
preshed==2.0.1
pylint==2.3.1
pymongo==3.7.2
pyparsing==2.4.7
python-dateutil==2.7.5
python-engineio==3.13.2
python-resize-image==1.1.19
python-socketio==4.6.0
pytz==2019.1
PyYAML==3.13
regex==2020.11.11
requests==2.21.0
singledispatch==3.4.0.3
six==1.12.0
spacy==2.1.3
SQLAlchemy==1.2.18
srsly==0.0.5
thinc==7.0.4
tqdm==4.31.1
typed-ast==1.4.1
urllib3==1.24.1
wasabi==0.2.1
Werkzeug==0.15.2
wincertstore==0.2
wrapt==1.11.1
zipp==3.4.0
